+++ src/org/bds/test/TestCasesZzz.java
@@ -1,12 +1,7 @@
 package org.bds.test;
 
-import java.io.File;
-
 import org.bds.Config;
-import org.bds.util.Gpr;
-import org.junit.Assert;
 import org.junit.Before;
-import org.junit.Test;
 
 /**
  * Quick test cases when creating a new feature...
@@ -20,39 +15,6 @@ public class TestCasesZzz extends TestCasesBase {
 	public void beforeEachTest() {
 		Config.reset();
 		Config.get().load();
-	}
-
-	@Test
-	public void test19_ftp_download() {
-		Gpr.debug("Test");
-		verbose = true;
-		String localFilePath = runAndGet("test/remote_19.bds", "fLocal").toString();
-
-		// Check that the file exists (remove tmp file after)
-		File f = new File(localFilePath);
-		Assert.assertTrue("Local file '" + localFilePath + "' does not exists", f.exists());
-		f.delete();
-	}
-
-	@Test
-	public void test20_ftp_exists() {
-		Gpr.debug("Test");
-		verbose = true;
-		runAndCheck("test/remote_20.bds", "fExists", "true");
-	}
-
-	@Test
-	public void test21_ftp_dir() {
-		Gpr.debug("Test");
-		verbose = true;
-		runAndCheck("test/remote_21.bds", "dHasReadme", "true");
-	}
-
-	@Test
-	public void test22_ftp_dir() {
-		Gpr.debug("Test");
-		verbose = true;
-		runAndCheck("test/remote_22.bds", "dHasReadme", "true");
 	}
 
 	//	@Test
