+++ test/z.bds
@@ -1,13 +1,17 @@
 #!/usr/bin/env bds
 
 # Test Ftp 1: Download file
+sleepTime := 5
 
-f := "ftp://ftp.ensembl.org/pub/release-96/gtf/homo_sapiens/zz/"
+f := "ftp://ftp.ensembl.org/pub/release-96/gtf/homo_sapiens/README"
 f_exists := f.exists()
 println "f_exists: $f_exists"
 
-#f_local := f.download()
-#println "f_local: '$f_local'"
+println('SLEEP')
+sleep(sleepTime)
+
+f_local := f.download()
+println "f_local: '$f_local'"
 
 #d := "ftp://ftp.ensembl.org/pub/release-96/gtf/homo_sapiens"
